% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitting.R
\name{splitting}
\alias{splitting}
\title{Training and Testing Partitioning}
\usage{
splitting(df, var)
}
\arguments{
\item{df}{A data frame containing the data to be split.}

\item{var}{A character string specifying the name of the binary response variable
(column) in the data frame. The values in this column should be `1` and `0`.}
}
\value{
A vector of indices representing a random sample of half of the rows
where the specified variable is `1` and half where it is `0`.
}
\description{
This function is intended for splitting a dataframe into a training set and a
testing set given a binary response variable. It creates a training set by
randomly sampling half of the instances where the specified variable has a
value of `1` and half of theinstances where it has a value of `0`.
The function returns the indices of the selected samples for further use.
}
\examples{
# Example usage with a data frame df and a binary column "target"
df <- data.frame(target = c(1, 0, 1, 0, 1, 1, 0, 0))
train_indices <- splitting(df, "target")
train <- df[train_indices,]
test <- df[-train_indices,]

}
