% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tjadasyn.R
\name{tjadasyn}
\alias{tjadasyn}
\title{Modifications to Adaptive Synthetic Sampling (ADASYN) for Imbalanced Data}
\usage{
tjadasyn(
  df,
  var,
  k = 5,
  over_ratio = 1,
  num_features = dim(df)[2] - 1,
  subsetFeat = FALSE,
  completeData = FALSE,
  corrupt = NULL,
  diminish = 0.5
)
}
\arguments{
\item{df}{A data frame containing the dataset.}

\item{var}{A character string specifying the name of the binary response variable in the data frame.}

\item{k}{An integer specifying the number of nearest neighbors to consider for the ADASYN algorithm. Default is 5.}

\item{over_ratio}{A numeric value for the oversampling ratio (the factor by which the minority class is oversampled). Default is 1.}

\item{num_features}{An integer specifying the number of features to consider. Defaults to all columns except the response variable.}

\item{subsetFeat}{A logical indicating whether to select the most informative features based on correlation with the response variable. Default is `FALSE`.}

\item{completeData}{A logical indicating whether to use the entire dataset (TRUE) or a subset of selected features (FALSE). Default is `FALSE`.}

\item{corrupt}{A numeric value specifying the percentage of data corruption to apply (e.g., random noise). Default is `NULL` (no corruption).}

\item{diminish}{A numeric value between 0 and 1 that controls the proportion features included based on correlation with the response variable. Default is 0.5.}
}
\value{
A data frame of synthetic samples generated by the ADASYN algorithm to balance the class distribution.
}
\description{
Improved Adaptive Synthetic Sampling (ADASYN) for balancing class distributions
in imbalanced datasets. It can optionally select informative features based
on correlation with the response variable and includes two optional tuning
parameters: corrupt and diminish. Corrupt will simulate addition of random noise
and diminish will reduce the feature space to eliminate junk variables.
}
\details{
The function first optionally selects a subset of features based on their correlation with the response variable (`var`). Then, the ADASYN algorithm is applied to create synthetic samples of the minority class. The oversampling ratio and number of nearest neighbors are parameters for the ADASYN process. Data corruption and feature selection are optional steps to enhance the generalizability of the model.
}
